backend/requirements.txt
# requirements.txt - Dependências do backend AnalisadorProduçãoWeb
# Siga esta lista para instalar todos os pacotes necessários para FastAPI, banco de dados, autenticação, análise e utilidades

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.0.3
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.0
scipy==1.11.4
plotly==5.17.0
loguru==0.7.2
slowapi==0.1.9